Some cleanups and fixes. Experimental algorithms.
In addition to cleanups this also adds a fast cluster-fit algorithm that precomputed the constant terms of the least squares problem. This is only useful when the number of unique colors per block is close to 16, with a small number of colors the weighted cluster fit algorithm performs better. However, this algorithm is much easier to parallelize (this is what the CUDA implementation did). This can be enabled with ICBC_FAST_CLUSTER_FIT.
There's also an experimental iterative cluster fit, ala k-means that never performed as well as I was hoping. This is exposed through compress_dxt1_test.
For a slight quality improvement you can also enable perfect rounding of color endpoints in the cluster fit algorithm with ICBC_PERFECT_ROUND, but this only works in the scalar code path.